style: moved vision-mission.tsx to container folder
