+++ src/app/about-us/page.tsx
@@ -1,6 +1,6 @@
 import PartnersView from '@/features/about-us/containers/partners-view';
 import About from '@/features/about-us/section/about';
-import VisionMission from '@/features/about-us/section/vision-mission';
+import VisionMission from '@/features/about-us/containers/vision-mission';
 
 export default function AboutUs() {
   return (
